+++ eslint.config.js
@@ -1,6 +1,5 @@
-import radham from './src/index.js';
+import radham, { radhamGfm } from './src/index.js';
 import json from '@eslint/json';
-import markdown from '@eslint/markdown';
 import { defineConfig } from 'eslint/config';
 import globals from 'globals';
 
@@ -18,8 +17,6 @@ export default defineConfig([
   },
   {
     files: ['**/*.md'],
-    plugins: { markdown },
-    extends: ['markdown/recommended'],
-    language: 'markdown/gfm'
+    extends: [radhamGfm]
   }
 ]);