+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+
+### Added
+
+- A `radhamGfm` config for [GitHub Flavored Markdown](https://github.github.com/gfm/)
+  documents.
+
 [8.0.0] - 2025-10-20
 --------------------
 
@@ -95,6 +103,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/jbenner-radham/eslint-config/compare/v8.0.0...HEAD
 [8.0.0]: https://github.com/jbenner-radham/eslint-config/compare/v7.0.0...v8.0.0
 [7.0.0]: https://github.com/jbenner-radham/eslint-config/compare/v6.0.0...v7.0.0
 [6.0.0]: https://github.com/jbenner-radham/eslint-config/compare/v5.0.0...v6.0.0